+++ packages/react-sample-library/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pantheon-systems/pcc-vue-sdk
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [a6c7134]
+  - @pantheon-systems/pcc-react-sdk@4.0.0
+  - @pantheon-systems/pcc-sdk-core@4.0.0
+
 ## 3.13.1
 
 ### Minor Changes